requirements.txt
djangorestframework~=3.16.0
pytz~=2025.2
numpy~=2.2.4
sqlalchemy~=2.0.40
Django~=5.1.7
alembic~=1.15.2
django-cors-headers==4.7.0
django-environ==0.12.0
